Initial setup of gems, and markdown syntax highlighting with line number
Took some inspiration from https://github.com/richardkall/richardkall.se
for the syntax highlighting theme. In addition, the initial middleman
setup with gems and the config are a derivative from that repository.

I've styled it such that code snippets have their line numbers, but are
fixed in location while the 'code' can scroll. Thus the line numbers
are always in view even if you have to scroll in the x-axis. We'll
see how this fairs in the long run.