Able to use bower, use bootstrap for its grid system
Use bootstrap and use its grid system to handle 'horizontal' sections
like the meta information for articles and the header navigation.

Bootstrap also allows us to handle different screen sizes in a more
responsive style. As Bootstrap's default CSS adds some styles to a bunch
of HTML elements, so I had to refactor a bunch of the SCSS to work with
bootstrap.