Update the feed.xml.builder to push publication datetime to end of day
The issue here is that the published time is based on the date only. For
newsletter (mailchimp) to pick up the change for daily pushes it needs
to have the published article to be within the last 24 hours. This
modification will push those publish dates near the end of the day.

Not the best solution, but it will do.

Also remove a ref link to the feed itself.