Use better Google Analytics (via philipwalton/analyticsjs-boilerplate)
Using https://github.com/philipwalton/analyticsjs-boilerplate to achieve
more with Google Analytics (mainly scroll distance, along with other
things).

On the admin I had to add the custom dimensions and metrics according to
how they are mentioned in this PR:
  https://github.com/philipwalton/analyticsjs-boilerplate/pull/15

The analyticsjs-boilerplate code uses webpacker and in a ways depends on
a build system to ensure imports from the autotrack library are hooked
in correctly. As this blog, in its current form, doesn't not have
anything like that. I decided to simply clone the repository and build
the project, producing the minified versions I need. I put these files
into `/vendor/analytics/`, which I hopefully never have to change again.

Add a script which tracks uncaught exceptions (used with the analytics
boilerplate trackers).

Remove the old adjusted bounce rate as the scroll metrics should do the
same (:fingerscrossed).