Update vendor/analytics (analyticsjs-boilerplate) for interactive scroll
Update the vendor code to include proper timezone (American/Toronto).

Update the vendor code to include a max scroll threshold (50%) to then
indicate that the session was 'interactive' thus having an adjusted
bounce rate. This means if they view one page, and navigate 50% of the
page they will not be counted as a bounce if they left the site.

Remove the source map references.