fix: feed.xml published and updated times to be unchanged
There are some odd situations where a post will be double sent.

I suspect that the publish time is pushed to end of day, so it is
possible for a post to go out twice (morning check send, then next
morning its published time is still past the last morning check).

So the idea is to just not change the time and leave it as is.

I've also modified Mailchimp's RSS send time to 11am so I can double
check to make sure things work as expected (8am was sometimes too early
for me to double check).